tool/source_code/src/updates/src/main.py
import gym
import openai
import ray
from ray import rllib
import zmq
import pika
import json

# Initialize the multi-agent AI system
def initialize_system():
    # Initialize OpenAI Gym environment
    env = gym.make('CartPole-v1')

    # Initialize Ray
    ray.init()

    # Initialize ZeroMQ context
    context = zmq.Context()

    # Initialize RabbitMQ connection
    connection = pika.BlockingConnection(pika.ConnectionParameters('localhost'))
    channel = connection.channel()

    return env, context, channel

# Integrate OpenAI Gym and GPT-3
def integrate_gym_gpt3(env):
    # Define the observation space, action space, and reward function
    observation_space = env.observation_space
    action_space = env.action_space

    # Define the agent's policy using GPT-3
    def agent_policy(observation):
        response = openai.Completion.create(
            engine="davinci-codex",
            prompt=f"Observation: {observation}\nAction:",
            max_tokens=10
        )
        action = int(response.choices[0].text.strip())
        return action

    return agent_policy

# Implement real-time communication protocols
def implement_communication_protocols(context, channel):
    # ZeroMQ communication
    socket = context.socket(zmq.PUB)
    socket.bind("tcp://*:5555")

    # RabbitMQ communication
    channel.queue_declare(queue='agent_queue')

    def send_message(message):
        # Send message via ZeroMQ
        socket.send_string(message)

        # Send message via RabbitMQ
        channel.basic_publish(exchange='',
                              routing_key='agent_queue',
                              body=message)

    return send_message

if __name__ == "__main__":
    env, context, channel = initialize_system()
    agent_policy = integrate_gym_gpt3(env)
    send_message = implement_communication_protocols(context, channel)

    # Example usage
    observation = env.reset()
    action = agent_policy(observation)
    send_message(json.dumps({"observation": observation.tolist(), "action": action}))
