tool/source_code/src/updates/src/utils/threat_intelligence.py
import requests
import json

class ThreatIntelligence:
    def __init__(self, platform_url):
        self.platform_url = platform_url

    def fetch_threat_data(self):
        response = requests.get(self.platform_url)
        if response.status_code == 200:
            return response.json()
        else:
            return None

    def process_threat_data(self, data):
        # Process and normalize threat data
        processed_data = json.dumps(data)
        return processed_data

    def update_attack_strategies(self, red_team_ai, threat_data):
        # Update red teaming AI's attack strategies based on new threat intelligence
        red_team_ai.update_strategies(threat_data)

if __name__ == "__main__":
    platform_url = "https://example.com/threat_intelligence"
    threat_intelligence = ThreatIntelligence(platform_url)
    threat_data = threat_intelligence.fetch_threat_data()
    if threat_data:
        processed_data = threat_intelligence.process_threat_data(threat_data)
        # Assuming red_team_ai is an instance of the red teaming AI class
        # threat_intelligence.update_attack_strategies(red_team_ai, processed_data)
